requirements.txt
flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Pillow==10.0.1
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
pytube==15.0.0
numpy==1.25.2
google-cloud-storage==2.10.0
pytz==2023.3
pyjwt==2.8.0
werkzeug==2.3.7
flask-cors==4.0.0
flask-jwt-extended==4.5.3